Resolved comments of 'Java Collections & Attendance updating | Miguel A Valdes G #66'
