+++ MiguelAValdesG/HackerRank/collections/hashtables/CheckMagazine.java
@@ -1,16 +1,10 @@
-package collections;
+package collections.hashtables;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Stream;
 
-import static java.util.stream.Collectors.toList;
-
-class Result {
+public class CheckMagazine {
 
   public static boolean checkMagazine(List<String> magazine, List<String> note) {
     Map<String, Integer> wordsInMagazine = new HashMap<>();
@@ -50,28 +44,3 @@ class Result {
     return CAN_BE_FORMED;
   }
 }
-
-public class JavaHashTablesRansomNote {
-
-  public static void main(String[] args) throws IOException {
-    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
-    final String CAN_BE_FORMED = "Yes";
-    final String CAN_NOT_BE_FORMED = "No";
-
-    List<String> magazine = Stream.of(bufferedReader.readLine().replaceAll("\\s+$", "").split(" "))
-      .collect(toList());
-
-    List<String> note = Stream.of(bufferedReader.readLine().replaceAll("\\s+$", "").split(" "))
-      .collect(toList());
-
-    if (Result.checkMagazine(magazine, note)) {
-      System.out.println(CAN_BE_FORMED);
-    } else {
-      System.out.println(CAN_NOT_BE_FORMED);
-    }
-
-    bufferedReader.close();
-  }
-}
-
-// Challenge: https://www.hackerrank.com/challenges/ctci-ransom-note/problem
