+++ MiguelAValdesG/HackerRank/collections/sort/JavaSort.java
@@ -1,32 +1,6 @@
-package collections;
+package collections.sort;
 
 import java.util.*;
-
-class Student {
-
-  private int id;
-  private String firstName;
-  private double cgpa;
-
-  public Student(int id, String firstName, double cgpa) {
-    super();
-    this.id = id;
-    this.firstName = firstName;
-    this.cgpa = cgpa;
-  }
-
-  public int getId() {
-    return id;
-  }
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public double getCgpa() {
-    return cgpa;
-  }
-}
 
 public class JavaSort {
