+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceI/Animal.java
@@ -1,0 +1,10 @@
+package fundamentals.inheritanceI;
+
+public class Animal {
+
+  void walk() {
+    final String WALKING = "I am walking";
+    System.out.println(WALKING);
+  }
+
+}
