+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceI/Bird.java
@@ -1,0 +1,15 @@
+package fundamentals.inheritanceI;
+
+public class Bird extends Animal {
+
+  void fly() {
+    final String FLYING = "I am flying";
+    System.out.println(FLYING);
+  }
+
+  void sing() {
+    final String SINGING = "I am singing";
+    System.out.println(SINGING);
+  }
+
+}
