+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceI/JavaInheritanceI.java
@@ -1,0 +1,12 @@
+package fundamentals.inheritanceI;
+
+public class JavaInheritanceI {
+
+    public static void main(String[] args) {
+      Bird bird = new Bird();
+      bird.walk();
+      bird.fly();
+      bird.sing();
+    }
+
+}
