+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceII/Adder.java
@@ -1,0 +1,9 @@
+package fundamentals.inheritanceII;
+
+public class Adder extends Arithmetic {
+
+  public int throwAdd(int firstNumber, int secondNumber) {
+    return add(firstNumber, secondNumber);
+  }
+
+}
