+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceII/Arithmetic.java
@@ -1,0 +1,9 @@
+package fundamentals.inheritanceII;
+
+public class Arithmetic {
+
+  public int add(int firstNumber, int secondNumber) {
+    return firstNumber + secondNumber;
+  }
+
+}
