+++ MiguelAValdesG/HackerRank/fundamentals/inheritanceII/JavaInheritanceII.java
@@ -1,0 +1,12 @@
+package fundamentals.inheritanceII;
+
+class JavaInheritanceII {
+
+  public static void main(String []args) {
+    Adder a = new Adder();
+
+    System.out.println("My superclass is: " + a.getClass().getSuperclass().getName());
+    System.out.print(a.add(10,32) + " " + a.add(10,3) + " " + a.add(10,10) + "\n");
+  }
+
+}
